Split E2E test files, prettier Travis jobs
